Modified the logic to be more complete

+++ lib/puppetx/simp/compliance_mapper.rb
@@ -336,25 +336,22 @@ def compiler_class()
                     end
 
                     fact_value = @callback.lookup_fact(confinement_setting)
-                    #require 'pry-byebug'; binding.pry() if confinement_value == ['--RedHat'] && fact_value == 'CentOS'
-                    if confinement_value.is_a?(Array)
-                      confined = false
-                      confinement_value.delete_if{ |confinement| !confinement.match(%r{^--}) && !fact_value.include?(confinement) }
-                      confinement_value.each do |confinement|
-                        if confinement.match(%r{^--})
-                          confined = true if fact_value.include?(confinement[2..-1])
+                    unless fact_value.nil?
+                      if confinement_value.is_a?(Array)
+                        confined = false
+                        confinement_value.delete_if{ |confinement| !confinement.match(%r{^--}) && !fact_value.include?(confinement) }
+                        confinement_value.map{ |confinement| confined = true if (confinement.match(%r{^--}) && fact_value.include?(confinement[2..-1])) }
+
+                        if confined || confinement_value.empty?
+                          delete_item = true
+                          throw :confine_end
+                        end
+                      else
+                        unless confinement_value.match(%r{^--}) ? (fact_value != confinement_value[2..-1]) : (fact_value == confinement_value)
+                          delete_item = true
+                          throw :confine_end
                         end
                       end
-
-                      if confined || confinement_value.empty?
-                        delete_item = true
-                        throw :confine_end
-                      end
-                    else
-                      unless confinement_value.match(%r{^--}) ? (fact_value != confinement_value[2..-1]) : (fact_value == confinement_value)
-                        delete_item = true
-                        throw :confine_end
-                      end  
                     end
                   end
                 end
